Move the event channel test helper to the test module.
Also suppresses some warnings, and organizes the test code with better packages and file names.
