+++ workflow-test/README.md
@@ -1,0 +1,3 @@
+# workflow-test
+
+This module contains utilities for testing workflow-based code.
